genai-usecases/agentic-ai/ai_workflow/requirements.txt
streamlit==1.41.1

pydantic==2.10.4
python-dotenv==1.0.1
plotly==5.24.1

langchain==0.3.14 
langchain-core==0.3.29
langgraph==0.2.61

groq==0.13.1 
google.generativeai==0.8.3
openai==1.59.4
anthropic==0.42.0